+++ README.md
@@ -1,10 +1,11 @@
-## 基于mediapipe进行手语识及翻译的研究
+## 基于mediapipe进行手语翻译的研究
 
 -----
 
 ## 技术栈
 
 -----
+开发语言:Python3.10
 
 模型训练:tensorflow
 
@@ -14,18 +15,20 @@
 
 通信协议:socket.io
 
-## 文件结构
+## 目录结构
 
 -----
 
-| 文件名/目录名              | 说明           |
-|----------------------|--------------|
-| bs-demo/static       | 静态资源css,js   |
-| bs-demo/templates    | h5           |
-| bs-demo/server.py    | flask server |
-| bs-demo/utils.py     | 工具类          |
-| other/localcamera.py | 本地摄像头测试      |
-| other/train.py       | 模型训练         |
-| other/utils.py       | 工具类          |
-| venv                 | Python虚拟环境   |
-| test.py              | 代码测试         |
+| 目录名/文件名&目录名         | 说明           |
+|---------------------|--------------|
+| prob/static         | 静态资源css,js   |
+| prob/templates      | h5           |
+| prob/server.py      | flask server |
+| prob/utils.py       | 工具类          |
+| test/localcamera.py | 本地摄像头测试      |
+| test/train.py       | 模型训练         |
+| test/utils.py       | 工具类          |
+| test/test.py        | 代码测试         |
+| train/              | 模型训练         |
+| venv                | Python虚拟环境   |
+|                     |              |
