--- bs-demo/gunicorn.conf
@@ -1,5 +1,0 @@
-bind = "0.0.0.0:5000"
-workers = 8
-threads = 8
-debug= True
-capture_output = True
