+++ prob/templates/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="ja">
+<head>
+    <title>Real-time Video Streaming</title>
+</head>
+<body>
+<video id="srcvideo" autoplay src></video>
+<img id="respimg" autoplay resp>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/4.6.1/socket.io.js"></script>
+<script src="/static/main.js"></script>
+</body>
+</html>
