+++ prob/uwsgi.ini
@@ -1,0 +1,14 @@
+[uwsgi]
+wsgi-file=/home/thinkzhao/PycharmProjects/sign-language-translate/bs-demo/server.py
+virtualenv = /home/thinkzhao/PycharmProjects/sign-language-translate/venv/
+callable=app
+http=0.0.0.0:5000
+thunder-lock=true
+master = true         
+processes=1
+threads = 4
+vhost = true          
+no-site = true        
+vacuum = true         
+memory-report=true
+buffer-size = 32768
